Move all test cases to ignore folder
